rename server.js file to app.js
